backend/foodgram/api/filters.py
from django_filters.rest_framework import FilterSet, filters
from recipes.models import Recipe, Tag
from rest_framework.filters import SearchFilter
from users.models import User


class IngredientFilter(SearchFilter):
    search_param = 'name'


class RecipeFilter(FilterSet):
    tags = filters.ModelMultipleChoiceFilter(
        queryset=Tag.objects.all(),
        to_field_name='slug',
        field_name='tags__slug',
        label='Tags'
    )
    author = filters.ModelChoiceFilter(queryset=User.objects.all())
    is_favorited = filters.BooleanFilter(
        method='get_is_favorited'
    )
    is_in_shopping_cart = filters.BooleanFilter(
        method='get_is_in_shopping_cart'
    )

    class Meta:
        model = Recipe
        fields = ('tags', 'author', 'is_favorited', 'is_in_shopping_cart')

    def get_is_favorited(self, queryset, name, data):
        if data:
            user = self.request.user
            if user.is_authenticated:
                return queryset.filter(favorite_recipe__author=user)
        return queryset

    def get_is_in_shopping_cart(self, queryset, name, data):
        if data:
            user = self.request.user
            if user.is_authenticated:
                return queryset.filter(shopping_list__user=user)
        return queryset
