backend/foodgram/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==3.2.16
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.0.4
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
iniconfig==1.1.1
isort==5.11.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==21.3
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.0.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2020.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
tomli==2.0.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.11.0
